Add initial empty hugo site
This is exactly what is created by `hugo new site xx` PLUS
we added one `.gitignore` file into every empty directory that
is created, so these empty-ISH directories are part of the repository
(because git doesn't let you add empty directories to your project).

that is, we want to be sure everybody all has the same empty-ish
directories, in case they are necessary for us to use hugo.